Fixed filenames from previous commit
Decouple phishing OSS suite from internal Splunk branding for community release

  - Replace hardcoded phishing inbox address with playbook_input:phishing_inbox_email
    across Acknowledgement Email, Phishing Attachment Filter, and Pre Processing and Dispatch
  - Fixed custom function references to updated custom function names submitted via PR#233
  - Change all playbook references from internal repo to community/
  - Change all custom function references from internal repo to community/
  - Added instructions to replace asset names in Notes for each playbook
