+++ [Phishing OSS] Indicator Blocks.json
@@ -4,7 +4,7 @@
     "category": "Response",
     "coa": {
         "data": {
-            "description": "This playbook blocks malicious indicators extracted from a phishing analysis. It takes indicators from an \"SAA Indicators\" artifact and applies blocking actions across multiple security tools.",
+            "description": "This playbook blocks malicious indicators extracted from a phishing analysis. It takes indicators from an \"SAA Indicators\" artifact and applies blocking actions across multiple security tools.\n\n",
             "edges": [
                 {
                     "id": "port_0_to_port_2",
@@ -48,7 +48,7 @@
                     "targetPort": "13_in"
                 }
             ],
-            "hash": "922b5e5abd2e6a03bea5154ded867a17fd4074cb",
+            "hash": "f37cc06ccd9bf9079c8af3813186745da73fe301",
             "nodes": {
                 "0": {
                     "data": {
@@ -64,7 +64,7 @@
                     "type": "start",
                     "warnings": {},
                     "x": 1000,
-                    "y": 419.9999999999994
+                    "y": 419.99999999999926
                 },
                 "1": {
                     "data": {
@@ -91,7 +91,7 @@
                         },
                         "connector": "Cisco Firepower",
                         "connectorConfigs": [
-                            "deep_firepower_dummy"
+                            "cisco_firepower"
                         ],
                         "connectorId": "F501CF42-46D4-4023-A156-208640865AA8",
                         "connectorVersion": "v1",
@@ -159,7 +159,7 @@
                         },
                         "connector": "Cisco ESA",
                         "connectorConfigs": [
-                            "deep_esa_dummytest"
+                            "cisco_esa"
                         ],
                         "connectorId": "7bf76da9-f31c-448b-906b-7c6de9d3cc70",
                         "connectorVersion": "v1",
@@ -297,7 +297,7 @@
                         },
                         "connector": "Cisco Umbrella",
                         "connectorConfigs": [
-                            "deep_umbrella_dummytest"
+                            "cisco_umbrella"
                         ],
                         "connectorId": "96f3e021-5396-42d0-97f4-4fab683e9adb",
                         "connectorVersion": "v1",
@@ -355,7 +355,7 @@
                     "y": 900
                 }
             },
-            "notes": "This playbook blocks malicious indicators extracted from a phishing analysis. It takes indicators from an \"SAA Indicators\" artifact and applies blocking actions across multiple security tools."
+            "notes": "This playbook blocks malicious indicators extracted from a phishing analysis. It takes indicators from an \"SAA Indicators\" artifact and applies blocking actions across multiple security tools.\n\nPlease replace with your assets for Cisco Umbrella, Firepower, ESA or corresponding equivalent apps for the corresponding actions."
         },
         "input_spec": null,
         "output_spec": null,
@@ -363,9 +363,9 @@
         "playbook_type": "automation",
         "python_version": "3.13",
         "schema": "5.0.24",
-        "version": "8.4.0.154"
+        "version": "8.5.0.248"
     },
-    "create_time": "2026-03-18T06:35:57.733196+00:00",
+    "create_time": "2026-06-30T15:52:19.665358+00:00",
     "draft_mode": false,
     "labels": [
         "*"
