+++ [Phishing OSS] SAA Enrich.json
@@ -85,7 +85,7 @@
                     "type": "start",
                     "warnings": {},
                     "x": 1280,
-                    "y": -100.00000000000023
+                    "y": -100.00000000000034
                 },
                 "1": {
                     "data": {
@@ -100,8 +100,8 @@
                     "id": "1",
                     "type": "end",
                     "warnings": {},
-                    "x": 980,
-                    "y": 980
+                    "x": 1280,
+                    "y": 1000
                 },
                 "20": {
                     "data": {
@@ -410,7 +410,7 @@
                         },
                         "connector": "Splunk Attack Analyzer Connector for Splunk SOAR",
                         "connectorConfigs": [
-                            "gso-saa_prod"
+                            "splunk_saa"
                         ],
                         "connectorId": "de681fee-c552-45bf-9212-827b1c7529f8",
                         "connectorVersion": "v1",
@@ -469,7 +469,7 @@
                         },
                         "connector": "Splunk Attack Analyzer Connector for Splunk SOAR",
                         "connectorConfigs": [
-                            "gso-saa_prod"
+                            "splunk_saa"
                         ],
                         "connectorId": "de681fee-c552-45bf-9212-827b1c7529f8",
                         "connectorVersion": "v1",
@@ -521,12 +521,12 @@
                     "y": 420
                 }
             },
-            "notes": "This playbook detonates suspicious email attachments in Splunk Attack Analyzer (SAA) sandbox  and extracts indicators for downstream analysis.",
+            "notes": "This playbook detonates suspicious email attachments in Splunk Attack Analyzer (SAA) sandbox  and extracts indicators for downstream analysis.\n\nPlease replace with your own SAA asset or equivalent app asset for the corresponding actions.",
             "origin": {
                 "playbook_id": 16127,
-                "playbook_name": "[Phishing Pond] SAA Enrich",
+                "playbook_name": "[Phishing OSS] SAA Enrich",
                 "playbook_repo_id": 12,
-                "playbook_repo_name": "sgs-soarcloud-gso-dev"
+                "playbook_repo_name": "community"
             }
         },
         "input_spec": null,
@@ -548,7 +548,7 @@
         "playbook_type": "automation",
         "python_version": "3.13",
         "schema": "5.0.24",
-        "version": "8.4.0.154"
+        "version": "8.5.0.248"
     },
     "create_time": "2026-03-17T14:17:10.117644+00:00",
     "draft_mode": false,
