+++ [Phishing OSS] Scope Campaign recipients.json
@@ -64,7 +64,7 @@
                     "type": "start",
                     "warnings": {},
                     "x": 19.999999999999986,
-                    "y": -6.750155989720952e-13
+                    "y": -7.389644451905042e-13
                 },
                 "1": {
                     "data": {
@@ -91,7 +91,7 @@
                         },
                         "connector": "Splunk",
                         "connectorConfigs": [
-                            "srv-rpa-rz-sgs-dmo_splunk-rz_core"
+                            "splunk"
                         ],
                         "connectorId": "91883aa8-9c81-470b-97a1-5d8f7995f560",
                         "connectorVersion": "v1",
@@ -244,12 +244,12 @@
                     "y": 340
                 }
             },
-            "notes": "This playbook finds all recipients who received a phishing campaign email by searching Email logs in Splunk.",
+            "notes": "This playbook finds all recipients who received a phishing campaign email by searching Email logs in Splunk.\n\nPlease replace with your own Splunk asset.",
             "origin": {
                 "playbook_id": 16268,
-                "playbook_name": "[Phishing Pond] Campaign recipients Output",
+                "playbook_name": "[Phishing OSS] Campaign recipients Output",
                 "playbook_repo_id": 12,
-                "playbook_repo_name": "sgs-soarcloud-gso-dev"
+                "playbook_repo_name": "community"
             }
         },
         "input_spec": [
@@ -282,7 +282,7 @@
         "playbook_type": "data",
         "python_version": "3.13",
         "schema": "5.0.24",
-        "version": "8.4.0.154"
+        "version": "8.5.0.248"
     },
     "create_time": "2026-03-17T13:32:46.736089+00:00",
     "draft_mode": false,
