+++ [Phishing OSS] TIP Addition.json
@@ -99,7 +99,7 @@
                     "type": "start",
                     "warnings": {},
                     "x": 1000,
-                    "y": 419.99999999999955
+                    "y": 419.99999999999943
                 },
                 "1": {
                     "data": {
@@ -128,7 +128,7 @@
                         },
                         "connector": "SecAuto ThreatConnect",
                         "connectorConfigs": [
-                            "tc_clone_1743724055270"
+                            "threat_connect"
                         ],
                         "connectorId": "7e872d12-915f-42fd-8ba2-d118a90f21de",
                         "connectorVersion": "v1",
@@ -259,7 +259,7 @@
                         },
                         "connector": "SecAuto ThreatConnect",
                         "connectorConfigs": [
-                            "tc_clone_1743724055270"
+                            "threat_connect"
                         ],
                         "connectorId": "7e872d12-915f-42fd-8ba2-d118a90f21de",
                         "connectorVersion": "v1",
@@ -327,7 +327,7 @@
                         },
                         "connector": "SecAuto ThreatConnect",
                         "connectorConfigs": [
-                            "tc_clone_1743724055270"
+                            "threat_connect"
                         ],
                         "connectorId": "7e872d12-915f-42fd-8ba2-d118a90f21de",
                         "connectorVersion": "v1",
@@ -397,7 +397,7 @@
                         },
                         "connector": "SecAuto ThreatConnect",
                         "connectorConfigs": [
-                            "tc_clone_1743724055270"
+                            "threat_connect"
                         ],
                         "connectorId": "7e872d12-915f-42fd-8ba2-d118a90f21de",
                         "connectorVersion": "v1",
@@ -457,7 +457,7 @@
                     "y": 980
                 }
             },
-            "notes": "This playbook exports phishing indicators to a Threat Intelligence Platform (TIP) for centralized storage and sharing across security tools."
+            "notes": "This playbook exports phishing indicators to a Threat Intelligence Platform (TIP) for centralized storage and sharing across security tools.\n\nPlease replace with asset of your Threat Intel Portal App in place of threat_connect"
         },
         "input_spec": null,
         "output_spec": null,
@@ -465,7 +465,7 @@
         "playbook_type": "automation",
         "python_version": "3.13",
         "schema": "5.0.24",
-        "version": "8.4.0.154"
+        "version": "8.5.0.248"
     },
     "create_time": "2026-03-17T13:30:58.619233+00:00",
     "draft_mode": false,
